task3-database-monitoring/requirements.txt
psycopg2-binary
prettytable